+++ TestSuite/docker_native_image_config.py
@@ -3,6 +3,7 @@ from typing import Dict, Optional
 
 from demisto_sdk.commands.common.constants import NATIVE_IMAGE_FILE_NAME
 from demisto_sdk.commands.common.handlers import JSON_Handler
+from TestSuite.test_tools import suite_join_path
 
 json = JSON_Handler()
 
@@ -19,8 +20,7 @@ class DockerNativeImageConfiguration:
             with open(self.path, 'w') as file:
                 json.dump(data, file)
         else:
-            from demisto_sdk.commands.common.native_image import load_native_image_config
-            from demisto_sdk.tests.constants_test import NATIVE_IMAGE_TEST_CONFIG_PATH
-
+            default_native_image_config_file_path = 'assets/default_docker_native_image_config.json'
             with open(self.path, 'w') as file:
-                json.dump(load_native_image_config(NATIVE_IMAGE_TEST_CONFIG_PATH), file, indent=4)
+                with open(suite_join_path(default_native_image_config_file_path), 'r') as native_image_conf_file:
+                    json.dump(json.load(native_image_conf_file), file)