+++ demisto_sdk/commands/content_graph/objects/integration.py
@@ -1,3 +1,4 @@
+import logging
 from typing import TYPE_CHECKING, List
 
 from demisto_sdk.commands.content_graph.objects.base_content import BaseContent
@@ -11,6 +12,8 @@ from pydantic import Field
 from demisto_sdk.commands.common.constants import MarketplaceVersions
 from demisto_sdk.commands.content_graph.common import ContentType, RelationshipType
 from demisto_sdk.commands.content_graph.objects.integration_script import IntegrationScript
+
+logger = logging.getLogger('demisto-sdk')
 
 
 class Command(BaseContent, content_type=ContentType.COMMAND):  # type: ignore[call-arg]
@@ -77,9 +80,11 @@ class Integration(IntegrationScript, content_type=ContentType.INTEGRATION):  # t
 
         if supported_native_images := self.get_supported_native_images(
             marketplace=marketplace,
-            native_image_config_file_path=kwargs.get('native_image_config_file_path'),
             ignore_native_image=kwargs.get('ignore_native_image') or False
         ):
+            logger.debug(
+                f'Adding the following native images {supported_native_images} to integration {self.object_id}'
+            )
             data['script']['nativeImage'] = supported_native_images
 
         return data