+++ demisto_sdk/commands/content_graph/objects/integration_script.py
@@ -5,7 +5,7 @@ from pydantic import Field
 
 from demisto_sdk.commands.common.constants import MarketplaceVersions
 from demisto_sdk.commands.common.handlers import YAML_Handler
-from demisto_sdk.commands.common.native_image import NativeImageSupportedVersions
+from demisto_sdk.commands.common.native_image import ScriptIntegrationSupportedNativeImages, file_to_native_image_config
 from demisto_sdk.commands.content_graph.objects.content_item import ContentItem
 from demisto_sdk.commands.prepare_content.integration_script_unifier import IntegrationScriptUnifier
 
@@ -32,13 +32,12 @@ class IntegrationScript(ContentItem):
     def get_supported_native_images(
         self,
         marketplace: MarketplaceVersions,
-        native_image_config_file_path: Optional[str] = None,
         ignore_native_image: bool = False
     ) -> List[str]:
         if marketplace == MarketplaceVersions.XSOAR and not ignore_native_image:
-            return NativeImageSupportedVersions(
+            return ScriptIntegrationSupportedNativeImages(
                 _id=self.object_id,
                 docker_image=self.docker_image,
-                native_image_config_file_path=native_image_config_file_path
+                native_image_config=file_to_native_image_config()
             ).get_supported_native_image_versions()
         return []